% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{activityItem}
\alias{activityItem}
\title{An activity item}
\usage{
activityItem(..., title = NULL, img = NULL, day = NULL,
  month = NULL, url = NULL)
}
\arguments{
\item{...}{item content}

\item{title}{item title}

\item{img}{img path or url}

\item{day}{day of publication}

\item{month}{month of publication}

\item{url}{external link}
}
\description{
An activity item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
