% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{activityList}
\alias{activityList}
\title{An activity list}
\usage{
activityList(...)
}
\arguments{
\item{...}{Slot for activityItem}
}
\description{
An activity list
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    box(
     title = "Activity List",
     activityList(
      lapply(X = 1:3, FUN = function(i) {
       activityItem(
        title = "Desmond Davison",
        img = paste0("https://image.flaticon.com/icons/svg/1087/108783", i,".svg"),
        day = 13,
        month = "june",
        url = "http://www.google.com",
        "Raw denim you probably haven't heard of them jean shorts Austin.
        Nesciunt tofu stumptown aliqua butcher retro keffiyeh
        dreamcatcher synth."
       )
      })
     )
    )
   )
  ),
  server <- function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
