% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{alert}
\alias{alert}
\title{An alert}
\usage{
alert(..., title = NULL, status = "primary", dismissible = TRUE,
  width = 3)
}
\arguments{
\item{...}{Alert text}

\item{title}{Alert title}

\item{status}{Alert status: "danger", "warning", "info", "success" or "primary"}

\item{dismissible}{Whether the alert is closable or not. TRUE by default.}

\item{width}{Alert width. 3 by default.}
}
\description{
An alert
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPageCustom(
   gentelellaBody(
    alert(
     status = "warning",
     title = "An alert",
     "Best check yo self,
     you're not looking too good."
    )
   )
  ),
  server <- function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
