% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{box}
\alias{box}
\title{A box to put dashboard elements in}
\usage{
box(..., width = 4, height = NULL, title = "Box title", subtitle = NULL,
  collapsible = TRUE, closable = FALSE, dropdownMenu = NULL)
}
\arguments{
\item{...}{elements to put in the box}

\item{width}{Box width}

\item{height}{Box height}

\item{title}{Box title}

\item{subtitle}{Box subtitle}

\item{collapsible}{Whether the box can be collapsed. TRUE by default.}

\item{closable}{Whether the box can be closed. FALSE by default.}

\item{dropdownMenu}{Dropdown menu: list of dropdown links. See example below.}
}
\description{
A box to put dashboard elements in
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPageCustom(
   gentelellaBody(
    box(
     "test",
     width = 4,
     height = NULL,
     title = "Box title",
     closable = TRUE,
     dropdownMenu = list(
      a(href = "http://www.google.com", "Test", target = "_blank"),
      a(href = "http://www.google.com", "Test", target = "_blank")
     )
    )
   )
  ),
  server = function(input, output, session) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
