% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{boxWidget}
\alias{boxWidget}
\title{A box widget container}
\usage{
boxWidget(..., title = NULL, width = NULL)
}
\arguments{
\item{...}{slot for any widget}

\item{title}{widget title}

\item{width}{widget width}
}
\description{
A box widget container
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPageCustom(
   gentelellaBody(
    box(
     width = 6,
     title = "Box Widget",
     fluidRow(
      column(
       width = 3,
       align = "center",
       sliderInput(
        "obs",
        "Number of observations:",
        min = 0,
        max = 1000,
        value = 500
       )
      ),
      column(
       width = 9,
       boxWidget(
        title = "Widget",
        plotOutput("distPlot")
       )
      )
     )
    )
   )
  ),
  server <- function(input, output) {
   output$distPlot <- renderPlot({
    hist(rnorm(input$obs))
   })
  }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
