% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{contactBox}
\alias{contactBox}
\title{A Contact box}
\usage{
contactBox(..., head_title = NULL, main_title = NULL, img = NULL,
  footer_left = NULL, footer_right = NULL, width = 4)
}
\arguments{
\item{...}{elements to put in the box. Do not put too large items.}

\item{head_title}{Box header title}

\item{main_title}{Box main title}

\item{img}{Any picture url or path}

\item{footer_left}{Footer left content}

\item{footer_right}{Footer right content}

\item{width}{Box width. 3 by default}
}
\description{
A Contact box
}
\note{
Does not work perfectly
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPageCustom(
   gentelellaBody(
    contactBox(
     head_title = "Digital Strategist",
     main_title = "Nicole Pearson",
     img = "https://image.flaticon.com/icons/svg/145/145859.svg",
     footer_left = "Some text",
     footer_right = pieChart(value = 45, id = "contactChart"),
     quickList(
      quickListItem(icon = "calendar-o", name = "Settings"),
      quickListItem(icon = "bars", name = "Subscription")
     )
    )
   )
  ),
  server = function(input, output, session) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
