% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{dashboard_box}
\alias{dashboard_box}
\title{A box to put dashboard elements in}
\usage{
dashboard_box(..., width = 4, height = NULL, box_title = "Box title",
  menuItems = list(a(class = "collapse-link", icon("chevron-up"))))
}
\arguments{
\item{...}{elements to put in the box}

\item{width}{Box width}

\item{height}{Box height}

\item{box_title}{Title above}

\item{menuItems}{A list of other things to appear in top menu}
}
\value{
A box to put elements in
}
\description{
A box to put dashboard elements in
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   dashboard_box(
     "test",
     width = 4,
     height = NULL,
     box_title = "Box title",
     menuItems = list(
       a(class = "collapse-link", icon("chevron-up")),
       list(
         a(href = "http://www.google.com", "Test", target = "_blank"),
         a(href = "#", "Test2")
       ),
       a(class = "close-link", icon("close"))
     )
   ),
   title_tag = "Shiny HTML Template",
   site_title = a(class="site_title", icon("eye"), span("Shiny HTML"))
  ),
  server = function(input, output, session) {}
 )
}

}
