% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{gentelellaAlert}
\alias{gentelellaAlert}
\title{A jumbotron}
\usage{
gentelellaAlert(..., title = NULL, status = "primary",
  dismissible = TRUE)
}
\arguments{
\item{...}{Alert text}

\item{title}{Alert title}

\item{status}{Alert status: "danger", "warning", "info", "success" or "primary"}

\item{dismissible}{Whether the alert is closable or not. TRUE by default.}
}
\description{
A jumbotron
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    gentelellaAlert(
     status = "warning",
     title = "An alert",
     "Best check yo self,
     you're not looking too good."
    )
   )
  ),
  server <- function(input, output) {}
 )
}

}
