% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaBody.R
\name{gentelellaBody}
\alias{gentelellaBody}
\title{Create a Gentelella dashboard body content}
\usage{
gentelellaBody(...)
}
\arguments{
\item{...}{Items to place in the body}
}
\description{
Create a Gentelella dashboard body content
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
