% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaFooter.R
\name{gentelellaFooter}
\alias{gentelellaFooter}
\title{Create a Gentelella dashboard footer}
\usage{
gentelellaFooter(leftText = "Mark Edmondson and David Granjon",
  rightText = "2018")
}
\arguments{
\item{leftText}{Any left text}

\item{rightText}{Any right text}
}
\description{
Create a Gentelella dashboard footer
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
