% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{gentelellaGauge}
\alias{gentelellaGauge}
\title{A simple circular diagram item}
\usage{
gentelellaGauge(value, height = 220, width = 220)
}
\arguments{
\item{value}{Item value}

\item{height}{Canvas height. 220 px by default.}

\item{width}{Canvas width. 220 px by default.}
}
\description{
A simple circular diagram item
}
