% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{gentelellaLabel}
\alias{gentelellaLabel}
\title{Create a label or badge}
\usage{
gentelellaLabel(name = NULL, status = "primary", position = NULL,
  mode = "label")
}
\arguments{
\item{name}{label name}

\item{status}{label status: "danger", "warning", "info", "success" or "primary"}

\item{position}{label position: NULL by default, "pull-right" or "pull-left". If mode is "badge", position is NULL}

\item{mode}{Either "label" or "badge". "label" by default.}
}
\description{
Create a label or badge
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    gentelellaLabel(name = "David", status = "warning", mode = "badge"),
    gentelellaLabel(name = "Mark", position = "pull-right"),
    gentelellaLabel(name = "Isabella", status = "danger", position = "pull-left")
   )
  ),
  server <- function(input, output) {}
 )
}

}
