% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaNavbar.R
\name{gentelellaNavbar}
\alias{gentelellaNavbar}
\title{Create a Gentelella dashboard navbar}
\usage{
gentelellaNavbar(..., navbarItems = NULL)
}
\arguments{
\item{...}{Any left UI elements}

\item{navbarItems}{Items placed on the right side}
}
\description{
Create a Gentelella dashboard navbar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
