% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaNavbar.R
\name{gentelellaNotif}
\alias{gentelellaNotif}
\title{Create a Gentelella notification menu}
\usage{
gentelellaNotif(..., id, icon = "envelope-o", status = "primary",
  expanded = FALSE)
}
\arguments{
\item{...}{slot for gentelellaNotifItem}

\item{id}{menu id. Must be unique.}

\item{icon}{menu icon}

\item{status}{menu status: "danger", "warning", "info", "success" or "primary"}

\item{expanded}{Whether the dropdown menu is open or not. FALSE by default.}
}
\description{
Create a Gentelella notification menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
