% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaNavbar.R
\name{gentelellaNotifItem}
\alias{gentelellaNotifItem}
\title{Create a Gentelella notification item}
\usage{
gentelellaNotifItem(..., title = NULL, date = NULL, img = NULL)
}
\arguments{
\item{...}{notification content}

\item{title}{notification title}

\item{date}{notification date}

\item{img}{image url or path}
}
\description{
Create a Gentelella notification item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
