% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaUI.R
\name{gentelellaPage}
\alias{gentelellaPage}
\title{render a gentellaPage in ui.R}
\usage{
gentelellaPage(..., title_tag = "Gentelella Shiny", site_title = a(class =
  "site_title", icon("paw"), span("Shiny HTML")), menuItems = NULL,
  footer = NULL, tracking = NULL)
}
\arguments{
\item{...}{ui elements that will be rendered into a fluidRow per entry}

\item{title_tag}{Title tag passed to headerBoilerPlate}

\item{site_title}{Website title passed to sideBarBoilerPlate}

\item{menuItems}{Sidebar menu items created with \link{sideBarElement}}

\item{footer}{Footer message}

\item{tracking}{If not NULL, pass in \link{trackingTags} to enable web analytics}
}
\value{
a function suitable for a ui.R page
}
\description{
Use tagList() on elements you want in the same row
}

