% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{gentelellaRibbonBox}
\alias{gentelellaRibbonBox}
\title{A box with a ribbon}
\usage{
gentelellaRibbonBox(..., width = 3, height = 390, ribbon_text = NULL,
  title = NULL)
}
\arguments{
\item{...}{elements to put in the box}

\item{width}{Box width}

\item{height}{Box height}

\item{ribbon_text}{Ribbon text}

\item{title}{Box title}
}
\description{
A box with a ribbon
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    gentelellaRibbonBox(
      ribbon_text = "30 \% Off",
      title = "Ribbon Box",
      "If you've decided to go in development mode and
      tweak all of this a bit, there are few things
      you should do.",
      gentelellaGauge(10)
     )
    )
  ),
  server = function(input, output, session) {}
 )
}

}
