% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaSidebar.R
\name{gentelellaSidebar}
\alias{gentelellaSidebar}
\title{Create a Gentelella dashboard sidebar}
\usage{
gentelellaSidebar(..., site_title = shiny::HTML(paste(shiny::icon("paw"),
  "Shiny Gentelella")), url = NULL, fixed = FALSE, footer = NULL)
}
\arguments{
\item{...}{Items to place in the sidebar such as \link{sidebarMenu}, \link{sidebarProfile}}

\item{site_title}{Dashboard title}

\item{url}{Link to external ressource}

\item{fixed}{Whether the sidebar is fixed. FALSE by default.}

\item{footer}{Sidebar footer content, if any.}
}
\description{
Create a Gentelella dashboard sidebar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
