% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaSidebar.R
\name{gentelellaSidebarMenu}
\alias{gentelellaSidebarMenu}
\title{Create a Gentelella dashboard sidebar menu}
\usage{
gentelellaSidebarMenu(...)
}
\arguments{
\item{...}{Items to place in the sidebar menu}
}
\description{
Create a Gentelella dashboard sidebar menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
