% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaSidebar.R
\name{gentelellaSidebarProfile}
\alias{gentelellaSidebarProfile}
\title{Create a Gentelella dashboard sidebar profile item}
\usage{
gentelellaSidebarProfile(img = NULL, name = NULL)
}
\arguments{
\item{img}{profile image}

\item{name}{user name}
}
\description{
Create a Gentelella dashboard sidebar profile item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
