% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{gentelellaStars}
\alias{gentelellaStars}
\title{A rating tag}
\usage{
gentelellaStars(value)
}
\arguments{
\item{value}{value between 0 and 5}
}
\description{
A rating tag
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    gentelellaStars(value = 4)
   )
  ),
  server <- function(input, output) {}
 )
}

}
