% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{gentelellaTabPanel}
\alias{gentelellaTabPanel}
\title{Create a gentelella tabPanel}
\usage{
gentelellaTabPanel(..., tabName, active = FALSE)
}
\arguments{
\item{...}{Tab content}

\item{tabName}{Tab name: it will be also passed as the id argument. Should be unique.}

\item{active}{Whether the tab is active or not. FALSE bu default.}
}
\description{
To be included in a gentelellaTabSetPanel
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
