% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{gentelellaTabSetPanel}
\alias{gentelellaTabSetPanel}
\title{Create a gentelella tabSetPanel}
\usage{
gentelellaTabSetPanel(..., id, right = FALSE)
}
\arguments{
\item{...}{Slot for gentelellaTabPanel.}

\item{id}{TabSetPanel id. Should be unique.}

\item{right}{If TabSetPanel start from the right side. FALSE by default.}
}
\description{
Create a gentelella tabSetPanel
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    gentelellaTabSetPanel(
     id = "tabset1",
     gentelellaTabPanel(
       tabName = "Home",
       active = TRUE,
       "Raw denim you probably haven't heard of
       them jean shorts Austin. Nesciunt tofu stumptown
       aliqua, retro synth master cleanse. Mustache
       cliche tempor, williamsburg carles vegan helvetica.
       Reprehenderit butcher retro keffiyeh dreamcatcher synth.
       Cosby sweater eu banh mi, qui irure terr."
     ),
     gentelellaTabPanel(
       tabName = "Profile",
       active = FALSE,
       sliderInput(
         "obs",
         "Number of observations:",
         min = 0,
         max = 1000,
         value = 500
       ),
       plotOutput("distPlot")
     )
    )
   )
  ),
  server = function(input, output, session) {
   output$distPlot <- renderPlot({
    hist(rnorm(input$obs))
   })
  }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
