% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_layouts.R
\name{graph_box}
\alias{graph_box}
\title{Graph box}
\usage{
graph_box(..., width = 12, boxtitle = "Impressive Title",
  subtitle = "sub-title", datepicker = dateRangeInput("datepicker_id",
  NULL))
}
\arguments{
\item{...}{Other elements to appear in the box such as graphs.}

\item{width}{width of box}

\item{boxtitle}{Title of box}

\item{subtitle}{Sub-title of box}

\item{datepicker}{a dateRangeInput or similar}
}
\value{
a box with a datepicker to put plots in
}
\description{
A box for graphs
}

