% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{jumbotron}
\alias{jumbotron}
\title{A jumbotron}
\usage{
jumbotron(..., title = NULL)
}
\arguments{
\item{...}{Any UI element or text}

\item{title}{jumbotron title}
}
\description{
A jumbotron
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPageCustom(
   gentelellaBody(
    jumbotron(
     title = "Hello, world!",
     "This is a simple hero unit, a simple jumbotron-style
     component for calling extra attention to featured
     content or information."
    )
   )
  ),
  server <- function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
