% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{label}
\alias{label}
\title{Create a label or badge}
\usage{
label(name = NULL, status = "primary", position = NULL, mode = "label")
}
\arguments{
\item{name}{label name}

\item{status}{label status: "danger", "warning", "info", "success" or "primary".
If mode is "badge" statuses are "red", "orange", "green", "blue", "purple".}

\item{position}{label position: NULL by default, "pull-right" or "pull-left".
If mode is "badge", position is NULL}

\item{mode}{Either "label" or "badge". "label" by default.}
}
\description{
Create a label or badge
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPageCustom(
   gentelellaBody(
    box(
     title = "Labels",
     label(name = "David", status = "warning", mode = "badge"),
     br(), br(), br(),
     label(name = "Mark", position = "pull-right"),
     label(name = "Isabella", status = "danger", position = "pull-left")
    )
   )
  ),
  server <- function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
