% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{pieChart}
\alias{pieChart}
\title{A simple circular diagram item}
\usage{
pieChart(id, value, height = 220, width = 220)
}
\arguments{
\item{id}{Unique id.}

\item{value}{Item value}

\item{height}{Canvas height. 220 px by default.}

\item{width}{Canvas width. 220 px by default.}
}
\description{
A simple circular diagram item
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    gentelellaRibbonBox(
      ribbon_text = "30 \% Off",
      title = "pieChart",
      "If you've decided to go in development mode and
      tweak all of this a bit, there are few things
      you should do.",
      pieChart(id = "chart1", value = 10),
      pieChart(id = "chart2", value = 20)
     )
    )
  ),
  server = function(input, output, session) {}
 )
}

}
