% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_box}
\alias{profile_box}
\title{Render the profile box}
\usage{
profile_box(profile_name, profile_url)
}
\arguments{
\item{profile_name}{The user name}

\item{profile_url}{The user profile image url}
}
\value{
profile box
}
\description{
Render the profile box
}

