% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_nav}
\alias{profile_nav}
\title{Render the top navbar profile}
\usage{
profile_nav(profile_name, profile_url, menu_items = NULL)
}
\arguments{
\item{profile_name}{The user name}

\item{profile_url}{The user profile image url}
}
\value{
profile top nav
}
\description{
Render the top navbar profile
}

