% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profileElements}
\alias{profileElements}
\title{profileElements}
\usage{
profileElements(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
Something
}
\description{
Shiny Module for use with \link{profileElementsUI}
}
\details{
Call via \code{shiny::callModule(profileElements, "your_id")}
}

