% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profileElementsUI}
\alias{profileElementsUI}
\title{profileElements UI}
\usage{
profileElementsUI(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Shiny UI
}
\description{
Shiny Module for use with \link{profileElements}
}

