% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_bars.R
\name{progress_stack}
\alias{progress_stack}
\title{Progress bars}
\usage{
progress_stack(values, display_totals = NULL, small = TRUE,
  text_pos = c("top", "side"), bar_class = "progress-bar bg-green")
}
\arguments{
\item{values}{A named vector of values}

\item{display_totals}{A named vector same length as values showing totals}

\item{small}{If there should be small bars or big bars}

\item{bar_class}{The class of the bars themselves}
}
\value{
a stack of progress bars
}
\description{
Progress bars are scaled from 0 to 100
}

