% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{progressBar}
\alias{progressBar}
\title{Progress bar}
\usage{
progressBar(value, side = "left", status = "primary",
  striped = FALSE, width = 6)
}
\arguments{
\item{value}{progress value}

\item{side}{From which side the bar comes: "left" or "right". "left" by default.}

\item{status}{progress status: "danger", "warning", "info", "success" or "primary".}

\item{striped}{Whether the progress bar is striped or not. FALSE by default.}

\item{width}{Progress bar width. Between 1 and 12.}
}
\description{
Progress bars are scaled from 0 to 100
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    box(
     width = 3,
     title = "Progress Bars",
     progressBar(
      20,
      side = "left",
      status = "primary",
      striped = FALSE
     ),
     progressBar(
      70,
      side = "right",
      status = "danger",
      striped = TRUE
     )
    )
   )
  ),
  server <- function(input, output) {}
 )
}

}
