% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{quickListItem}
\alias{quickListItem}
\title{A quick list item}
\usage{
quickListItem(icon, name = NULL)
}
\arguments{
\item{icon}{item icon}

\item{name}{item name}
}
\description{
A quick list item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
