% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{ribbonBox}
\alias{ribbonBox}
\title{A box with a ribbon}
\usage{
ribbonBox(..., width = 3, height = 390, ribbon_text = NULL,
  title = NULL, ribbon_color = NULL)
}
\arguments{
\item{...}{elements to put in the box}

\item{width}{Box width}

\item{height}{Box height}

\item{ribbon_text}{Ribbon text}

\item{title}{Box title}

\item{ribbon_color}{Ribbon background color: "red", "orange", "blue", "purple", "green".}
}
\description{
A box with a ribbon
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    ribbonBox(
      ribbon_text = "30 \% Off",
      title = "Ribbon Box",
      "If you've decided to go in development mode and
      tweak all of this a bit, there are few things
      you should do.",
      pieChart(id = "chart1", value = 10),
      pieChart(id = "chart2", value = 20)
     )
    )
  ),
  server = function(input, output, session) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
