% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaSidebar.R
\name{sidebarDate}
\alias{sidebarDate}
\title{Create a Gentelella dashboard sidebar date}
\usage{
sidebarDate()
}
\description{
Create a Gentelella dashboard sidebar date
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
