% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplates.R
\name{sideBarElement}
\alias{sideBarElement}
\title{sideBarLinks}
\usage{
sideBarElement(element, icon = NULL, nested_element = NULL)
}
\arguments{
\item{element}{Element.}

\item{icon}{Icon.}

\item{nested_element}{Nested items if any.}
}
\description{
Things put in li()
}
\details{
Use HTML() wrapped around icon() if neccessary
}
