% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaSidebar.R
\name{sidebarItem}
\alias{sidebarItem}
\title{Create a Gentelella dashboard sidebar menu item}
\usage{
sidebarItem(tabName = NULL, icon = NULL)
}
\arguments{
\item{tabName}{item name}

\item{icon}{item icon}

\item{nested_element}{nested_element, if any}
}
\description{
Create a Gentelella dashboard sidebar menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
