% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaSidebar.R
\name{sidebarMenu}
\alias{sidebarMenu}
\title{Create a Gentelella dashboard sidebar menu}
\usage{
sidebarMenu(..., title = NULL)
}
\arguments{
\item{...}{Slot for \link{sidebarItem}}

\item{title}{Menu section title}
}
\description{
Create a Gentelella dashboard sidebar menu
}
\author{
Mark Edmondson, \email{m@sunholo.com}
}
