% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{socialBox}
\alias{socialBox}
\title{A social box to present a product or people}
\usage{
socialBox(..., width = 3, height = 390, title = NULL, url_1 = NULL,
  url_2 = NULL, media_1 = NULL, media_2 = NULL, profile_img = NULL,
  footer = NULL)
}
\arguments{
\item{...}{elements to put in the box}

\item{width}{Box width}

\item{height}{Box height}

\item{title}{Box title}

\item{url_1}{Box link 1}

\item{url_2}{Box link 2,}

\item{media_1}{Media name like twitter, facebook, ...}

\item{media_2}{Media name like twitter, facebook, ...}

\item{profile_img}{Box profile image}

\item{footer}{Box footer content, if any.}
}
\description{
A social box to present a product or people
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    socialBox(
     title = "Social Box",
     url_1 = "https://www.facebook.com",
     url_2 = "https://twitter.com",
     media_1 = "facebook",
     media_2 = "twitter",
     profile_img = "https://image.flaticon.com/icons/svg/17/17004.svg",
     footer = "If you\\'ve decided to go in development
     mode and tweak all of this a bit, there are few
     things you should do.",
     socialStats(
      socialStatsItem(value = 123, name = "Articles"),
      socialStatsItem(value = 1234, name = "Followers"),
      socialStatsItem(value = 435, name = "Following")
     )
    )
   )
  ),
  server = function(input, output, session) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
