% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{socialStatsItem}
\alias{socialStatsItem}
\title{A social stats item to insert in a socialStats container}
\usage{
socialStatsItem(value = NULL, name = NULL)
}
\arguments{
\item{value}{Item value}

\item{name}{Item name}
}
\description{
A social stats item to insert in a socialStats container
}
