% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaBody.R
\name{tabItems}
\alias{tabItems}
\title{A container for tab items}
\usage{
tabItems(...)
}
\arguments{
\item{...}{Items to put in the container. Each item should be a \code{\link{tabItem}}.}
}
\description{
A container for tab items
}
