% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabPill}
\alias{tabPill}
\title{Create a gentelella tabPill}
\usage{
tabPill(..., tabName, active = FALSE)
}
\arguments{
\item{...}{Tab content}

\item{tabName}{Tab name: it will be also passed as the id argument. Should be unique.}

\item{active}{Whether the tab is active or not. FALSE bu default.}
}
\description{
To be included in a \link{tabSetPill}
}
\note{
tabName must be unique even between \link{tabSetPanel} or \link{tabSetPill}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
