% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabSetPill}
\alias{tabSetPill}
\title{Create a gentelella tabSetPill}
\usage{
tabSetPill(..., right = FALSE)
}
\arguments{
\item{...}{Slot for \link{tabPill}.}

\item{right}{If TabSetPanel start from the right side. FALSE by default.}
}
\description{
Create a gentelella tabSetPill
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    tabSetPill(
     tabPill(
       tabName = "Home",
       active = TRUE,
       "Raw denim you probably haven't heard of
       them jean shorts Austin. Nesciunt tofu stumptown
       aliqua, retro synth master cleanse. Mustache
       cliche tempor, williamsburg carles vegan helvetica.
       Reprehenderit butcher retro keffiyeh dreamcatcher synth.
       Cosby sweater eu banh mi, qui irure terr."
     ),
     tabPill(
       tabName = "Profile",
       active = FALSE,
       sliderInput(
         "obs",
         "Number of observations:",
         min = 0,
         max = 1000,
         value = 500
       ),
       plotOutput("distPlot")
     )
    )
   )
  ),
  server = function(input, output, session) {
   output$distPlot <- renderPlot({
    hist(rnorm(input$obs))
   })
  }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
