% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tiles.R
\name{tileCountElement}
\alias{tileCountElement}
\title{tileCountRow Element}
\usage{
tileCountElement(value = 2500, change_value = "4\%", going_well = TRUE,
  tile_title = " Total Users", width = 2, icon_in = icon("user"),
  from_text = " From last Week", highlight = NULL)
}
\arguments{
\item{value}{Count value}

\item{change_value}{Change value}

\item{going_well}{If TRUE then change_value is green, else red}

\item{tile_title}{Title text}

\item{width}{Width of tile in bootstrap}

\item{icon_in}{Icon to show}

\item{from_text}{Change text}

\item{highlight}{color to highlight value}
}
\value{
a tileCountRow for use within \link{tileCountRow}
}
\description{
tileCountRow Element
}

