% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{tileCountRow}
\alias{tileCountRow}
\title{tileCountRow}
\usage{
tileCountRow(...)
}
\arguments{
\item{...}{tileCountElements}
}
\description{
tileCountRow
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    tileCountRow(
      lapply(1:4, tileCountElement)
    )
   )
  ),
  server <- function(input, output) {}
 )
}

}
\author{
Mark Edmondson, \email{m@sunholo.com}
}
