% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{timeline}
\alias{timeline}
\title{A timeline block}
\usage{
timeline(...)
}
\arguments{
\item{...}{slot for \link{timelineItem}}
}
\description{
A timeline block
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    gentelellaBox(
     width = 4,
     title = "Timeline",
     menuItems = list(
       a(class = "collapse-link", icon("chevron-up")),
       list(
         a(href = "http://www.google.com", "Test", target = "_blank"),
         a(href = "#", "Test2")
       ),
       a(class = "close-link", icon("close"))
     ),
     timeline(
      timelineItem(
       title = "Who Needs Sundance When You’ve Got Crowdfunding?",
       url = NULL,
       date = "13 hours ago",
       author = "Jane Smith",
       "Film festivals used to be do-or-die moments for movie makers.
       They were where you met the producers that could fund your
       project, and if the buyers liked your flick, they’d pay to
       Fast-forward and ..."
      ),
      timelineItem(
       title = "Who needs Money",
       url = "http:://www.google.com",
       date = "Today",
       author = "John Doe",
       "Nobody need money!",
       tag = "My tag"
      )
     )
    )
   )
  ),
  server = function(input, output, session) {}
 )
}

}
