% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{timelineItem}
\alias{timelineItem}
\title{A timeline item}
\usage{
timelineItem(..., title = NULL, url = NULL, date = NULL,
  author = NULL, tag = NULL)
}
\arguments{
\item{...}{timeline item content, any element}

\item{title}{timeline item title}

\item{url}{timelime item external link}

\item{date}{timeline item date}

\item{author}{timeline item author, if any}

\item{tag}{timeline item tag}
}
\description{
A timeline item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
