% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{trackingTags}
\alias{trackingTags}
\title{Tracking tags for your app}
\usage{
trackingTags(tag_template = c("ga", "gtm", "freeform"), tag_code)
}
\arguments{
\item{tag_template}{Tracking tag template}

\item{tag_code}{If GA, the UA code.  If GTM, the GTM code.

tag_code depends on what tag_template is:

\itemize{
  \item ga: The UA code (UA-xxxxxx-x)
  \item gtm The GTM code (GTM-xxxxx)
  \item freeform The JS code
 }
If you select a tag template, you can put UA-xxxx-x as tag_code}
}
\value{
JavaScript for tracking
}
\description{
Tracking tags for your app
}

