% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{userList}
\alias{userList}
\title{A user list block}
\usage{
userList(...)
}
\arguments{
\item{...}{slot for userListItem}
}
\description{
A user list block
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    gentelellaBox(
     width = 3,
     title = "user List",
     menuItems = list(
       a(class = "collapse-link", icon("chevron-up")),
       list(
         a(href = "http://www.google.com", "Test", target = "_blank"),
         a(href = "#", "Test2")
       ),
       a(class = "close-link", icon("close"))
     ),
     userList(
      userListItem(
       user_img = NULL,
       user_url = "http:://www.google.com",
       title = "user 1",
       footer = "2 Sales Today",
       "$2300. Agent Avarage Sales."
      )
     )
    )
   )
  ),
  server <- function(input, output) {}
 )
}

}
