% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{userListItem}
\alias{userListItem}
\title{A user list item}
\usage{
userListItem(..., user_img = NULL, title = NULL, footer = NULL)
}
\arguments{
\item{...}{Any content}

\item{user_img}{User image}

\item{title}{item title}

\item{footer}{footer content}
}
\description{
A user list item
}
