% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{valueBox}
\alias{valueBox}
\title{A value box}
\usage{
valueBox(value, title = NULL, description = NULL, icon = NULL,
  width = 3)
}
\arguments{
\item{value}{value}

\item{title}{value box title}

\item{description}{description if any}

\item{icon}{value box icon if any}

\item{width}{value box width. 3 by default}
}
\description{
A value box
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPageCustom(
   gentelellaBody(
    valueBox(
     value = 179,
     title = "New Sign ups",
     description = "Lorem ipsum psdea itgum rixt",
     icon = icon("caret-square-o-right")
    ),
    valueBox(
     value = 345,
     title = "New Customers",
     description = "Lorem ipsum psdea itgum rixt",
     icon = icon("comments-o")
    )
   )
  ),
  server = function(input, output, session) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
