% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{wizard}
\alias{wizard}
\title{A wizard container}
\usage{
wizard(..., orientation = "horizontal")
}
\arguments{
\item{...}{slot for wizardItem}

\item{orientation}{wizard orientation: "horizontal" or "vertical"}
}
\description{
A wizard container
}
