% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{wizardItem}
\alias{wizardItem}
\title{A wizard item}
\usage{
wizardItem(..., display = "none", description = NULL)
}
\arguments{
\item{...}{any UI element}

\item{display}{Whether to diplay if or not. "none" by default.}

\item{description}{Item description, if any.}
}
\description{
A wizard item
}
