% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_layouts.R
\name{dashboard_box}
\alias{dashboard_box}
\title{A box to put dashboard elements in}
\usage{
dashboard_box(..., width = 4, height = 320, box_title = "Box title",
  menuItems = list(a(class = "collapse-link", icon("chevron-up"))))
}
\arguments{
\item{...}{elements to put in the box}

\item{width}{Width}

\item{box_title}{Title above}

\item{menuItems}{A list of other things to appear in top menu}
}
\value{
A box to put elements in
}
\description{
Its 320 pixels high by default
}

