% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{gentelellaBox}
\alias{gentelellaBox}
\title{A box to put dashboard elements in}
\usage{
gentelellaBox(..., width = 4, height = NULL, title = "Box title",
  subtitle = NULL, menuItems = list(shiny::a(class = "collapse-link",
  shiny::icon("chevron-up"))))
}
\arguments{
\item{...}{elements to put in the box}

\item{width}{Box width}

\item{height}{Box height}

\item{title}{Box title}

\item{subtitle}{Box subtitle}

\item{menuItems}{A list of other things to appear in top menu such as close button, collapse and dropdowns}
}
\description{
A box to put dashboard elements in
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    gentelellaBox(
     "test",
     width = 4,
     height = NULL,
     title = "Box title",
     menuItems = list(
       a(class = "collapse-link", icon("chevron-up")),
       list(
         a(href = "http://www.google.com", "Test", target = "_blank"),
         a(href = "#", "Test2")
       ),
       a(class = "close-link", icon("close"))
     )
    )
   )
  ),
  server = function(input, output, session) {}
 )
}

}
