% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaSidebar.R
\name{gentelellaSidebar}
\alias{gentelellaSidebar}
\title{Create a Gentelella dashboard sidebar}
\usage{
gentelellaSidebar(..., site_title = HTML(paste(icon("paw"),
  "Shiny Gentelella")), url = NULL)
}
\arguments{
\item{...}{Items to place in the sidebar such as gentelellaSidebarMenu, gentelellaSidebarProfile}

\item{site_title}{Dashboard title}

\item{url}{Link to external ressource}
}
\description{
Create a Gentelella dashboard sidebar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
