% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaSidebar.R
\name{gentelellaSidebarItem}
\alias{gentelellaSidebarItem}
\title{Create a Gentelella dashboard sidebar menu item}
\usage{
gentelellaSidebarItem(tabName = NULL, icon = NULL,
  nested_element = NULL)
}
\arguments{
\item{icon}{item icon}
}
\description{
Create a Gentelella dashboard sidebar menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
