% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaNavbar.R
\name{notif}
\alias{notif}
\title{Create a Gentelella notification menu}
\usage{
notif(..., id, icon = icon("envelope-o"), status = "primary",
  expanded = FALSE)
}
\arguments{
\item{...}{slot for \link{notifItem}}

\item{id}{menu id. Must be unique.}

\item{icon}{menu icon}

\item{status}{menu status: "danger", "warning", "info", "success" or "primary"}

\item{expanded}{Whether the dropdown menu is open or not. FALSE by default.}
}
\description{
Create a Gentelella notification menu
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
   ui = gentelellaPageCustom(
    navbar = gentelellaNavbar(
    navbarItems = notif(
      id = "menunotif",
      icon = icon("envelope-o"),
      status = "danger",
      expanded = TRUE,
      lapply(X = 1:5, FUN = function(i) {
        notifItem(
          title = "John Doe",
          date = "3 min ago",
          img = paste0("https://image.flaticon.com/icons/svg/163/16382", i,".svg"),
          "Film festivals used to be do-or-die moments
          for movie makers. They were where..."
        )
      })
     )
    ),
    sidebar = gentelellaSidebar(),
    body = gentelellaBody(),
    footer = gentelellaFooter()
   ),
   server <- function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
