% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{progressBar}
\alias{progressBar}
\title{Progress bar}
\usage{
progressBar(value, side = "left", status = "primary",
  striped = FALSE, width = 6)
}
\arguments{
\item{value}{progress value}

\item{side}{From which side the bar comes: "left" or "right". "left" by default.}

\item{status}{progress status: "danger", "warning", "info", "success".}

\item{striped}{Whether the progress bar is striped or not. FALSE by default.}
}
\description{
Progress bars are scaled from 0 to 100
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    gentelellaBox(
     width = 3,
     title = "Progress Bars",
     menuItems = list(
       a(class = "collapse-link", icon("chevron-up")),
       list(
         a(href = "http://www.google.com", "Test", target = "_blank"),
         a(href = "#", "Test2")
       ),
       a(class = "close-link", icon("close"))
     ),
     progressBar(
      20,
      side = "left",
      status = "primary",
      striped = FALSE
     ),
     progressBar(
      70,
      side = "right",
      status = "danger",
      striped = TRUE
     )
    )
   )
  ),
  server <- function(input, output) {}
 )
}

}
