% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{quickList}
\alias{quickList}
\title{A quick list container}
\usage{
quickList(...)
}
\arguments{
\item{...}{slot for quickListItem}
}
\description{
A quick list container
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    gentelellaBox(
     width = 4,
     title = "Quick Lists",
     menuItems = list(
       a(class = "collapse-link", icon("chevron-up")),
       list(
         a(href = "http://www.google.com", "Test", target = "_blank"),
         a(href = "#", "Test2")
       ),
       a(class = "close-link", icon("close"))
     ),
     quickList(
      quickListItem(icon = "calendar-o", name = "Settings"),
      quickListItem(icon = "bars", name = "Subscription")
     )
    )
   )
  ),
  server = function(input, output, session) {}
 )
}

}
