% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{sideBarElement}
\alias{sideBarElement}
\title{sideBarLinks}
\usage{
sideBarElement(element, icon = NULL, nested_element = NULL)
}
\description{
Things put in li()
}
\details{
Use HTML() wrapped around icon() if neccessary
}

