% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaSidebar.R
\name{sidebarItem}
\alias{sidebarItem}
\title{Create a Gentelella dashboard sidebar menu item}
\usage{
sidebarItem(..., tabName = NULL, icon = NULL, badgeName = NULL,
  badgeStatus = "danger")
}
\arguments{
\item{...}{item name}

\item{tabName}{item id. Must be unique}

\item{icon}{item icon}

\item{badgeName}{item badge name if any.}

\item{badgeStatus}{item badge status: "danger", "warning", "info", "success" or "primary".}
}
\description{
Create a Gentelella dashboard sidebar menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
