% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentelellaSidebar.R
\name{sidebarItem}
\alias{sidebarItem}
\title{Create a Gentelella dashboard sidebar menu item}
\usage{
sidebarItem(..., tabName = NULL, icon = NULL)
}
\arguments{
\item{...}{item name}

\item{tabName}{item id. Must be unique}

\item{icon}{item icon}
}
\description{
Create a Gentelella dashboard sidebar menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
