% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tiles.R
\name{updateTileCount}
\alias{updateTileCount}
\title{updateTileCount}
\usage{
updateTileCount(input, output, session, value, change_value, going_well,
  tile_title = " Total Users", width = 2, icon_in = icon("user"),
  from_text = " From last Week", highlight = reactive(NULL))
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{value}{[reactive] Count value}

\item{change_value}{[reactive] Change value}

\item{going_well}{[reactive] If TRUE then change_value is green, else red}

\item{tile_title}{Title text}

\item{width}{Width of tile in bootstrap}

\item{icon_in}{Icon to show}

\item{from_text}{Change text}

\item{highlight}{[reactive] color to highlight value}
}
\description{
Shiny Module for use with \link{tileCountUI}
}
\details{
Call via \code{shiny::callModule(updateTileCount, "your_id")}
}

