% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{userList}
\alias{userList}
\title{A user list block}
\usage{
userList(...)
}
\arguments{
\item{...}{slot for userListItem}
}
\description{
A user list block
}
\examples{
if (interactive()) {
 library(shiny)
 library(gentelellaShiny)
 shinyApp(
  ui = gentelellaPage(
   gentelellaBody(
    box(
     width = 3,
     title = "user List",
     userList(
      userListItem(
       user_img = "https://image.flaticon.com/icons/svg/145/145862.svg",
       user_url = "http:://www.google.com",
       title = "user 1",
       subtitle = "2 Sales Today",
       "$2300. Agent Avarage Sales."
      ),
      userListItem(
       user_img = "https://image.flaticon.com/icons/svg/145/145864.svg",
       user_url = "http:://www.google.com",
       title = "user 2",
       subtitle = "4 Sales Today",
       "$4600. Agent Avarage Sales."
      )
     )
    )
   )
  ),
  server <- function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
